# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateMissionRequest": {
        "additionalProperties": false,
        "properties": {
          "constraints": {
            "description": "Operational constraints.",
            "items": {
              "type": "string"
            },
            "title": "Constraints",
            "type": "array"
          },
          "deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Completion deadline.",
            "title": "Deadline"
          },
          "goal": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Clear objective or goal of the mission.",
            "title": "Goal"
          },
          "kind": {
            "$ref": "#/components/schemas/MissionKind",
            "default": "general_goal",
            "description": "Classification of the mission."
          },
          "success_criteria": {
            "description": "Target outcomes.",
            "items": {
              "type": "string"
            },
            "title": "Success Criteria",
            "type": "array"
          },
          "title": {
            "description": "Title of the mission.",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "CreateMissionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MissionKind": {
        "description": "High-level classification of user objective.",
        "enum": [
          "opportunity_pursuit",
          "problem_resolution",
          "general_goal"
        ],
        "title": "MissionKind",
        "type": "string"
      },
      "MissionResponse": {
        "properties": {
          "archived": {
            "default": false,
            "title": "Archived",
            "type": "boolean"
          },
          "archived_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "title": "Constraints",
            "type": "array"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "goal": {
            "title": "Goal",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "mission_id": {
            "title": "Mission Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "success_criteria": {
            "items": {
              "type": "string"
            },
            "title": "Success Criteria",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "mission_id",
          "user_id",
          "title",
          "goal",
          "kind",
          "status",
          "created_at",
          "updated_at",
          "version"
        ],
        "title": "MissionResponse",
        "type": "object"
      },
      "MissionStatus": {
        "description": "Lifecycle states for a Mission.",
        "enum": [
          "draft",
          "active",
          "paused",
          "completed",
          "cancelled",
          "failed"
        ],
        "title": "MissionStatus",
        "type": "string"
      },
...
      "PaginatedResponse_MissionResponse_": {
        "properties": {
          "items": {
            "description": "Collection items for current page.",
            "items": {
              "$ref": "#/components/schemas/MissionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "description": "Current page size limit.",
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "description": "Offset position.",
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total number of items matching filter.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[MissionResponse]",
        "type": "object"
      },
...
      "TransitionMissionRequest": {
        "additionalProperties": false,
        "properties": {
          "new_status": {
            "$ref": "#/components/schemas/MissionStatus",
            "description": "Target lifecycle status."
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reason for the status transition.",
            "title": "Reason"
          }
        },
        "required": [
          "new_status"
        ],
        "title": "TransitionMissionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Autonomous caseworker agent service for proactive opportunities, cases, and actions.",
    "title": "JackVerse Caseworker API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/missions": {
      "get": {
        "description": "List missions belonging to the authenticated user with offset pagination.",
        "operationId": "list_missions_api_v1_missions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MissionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Archived",
              "type": "boolean"
            }
          },
          {
            "description": "Items per page (max 100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Items per page (max 100)",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Pagination offset",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Pagination offset",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_MissionResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Missions",
        "tags": [
          "missions"
        ]
      },
      "post": {
        "description": "Create a new user mission, binding ownership strictly to the authenticated Principal.",
        "operationId": "create_mission_api_v1_missions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Mission",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{mission_id}": {
      "get": {
        "description": "Retrieve mission ensuring strict cross-user isolation (404 on unowned missions).",
        "operationId": "get_mission_api_v1_missions__mission_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Mission",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{mission_id}/archive": {
      "post": {
        "description": "Archive a non-active mission. Idempotent. Requires valid ETag if provided.",
        "operationId": "archive_mission_api_v1_missions__mission_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Mission",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{mission_id}/cancel": {
      "post": {
        "description": "Convenience endpoint to cancel a mission.",
        "operationId": "cancel_mission_api_v1_missions__mission_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reason",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Mission",
        "tags": [
          "missions"
        ]
      }
    },
...
    "/api/v1/missions/{mission_id}/pause": {
      "post": {
        "description": "Convenience endpoint to pause an active mission.",
        "operationId": "pause_mission_api_v1_missions__mission_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reason",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Mission",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{mission_id}/restore": {
      "post": {
        "description": "Restore an archived mission. Idempotent. Requires valid ETag if provided.",
        "operationId": "restore_mission_api_v1_missions__mission_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Mission",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{mission_id}/resume": {
      "post": {
        "description": "Convenience endpoint to resume a paused mission.",
        "operationId": "resume_mission_api_v1_missions__mission_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reason",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Mission",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{mission_id}/transition": {
      "post": {
        "description": "Transition mission status with strict ETag precondition verification.",
        "operationId": "transition_mission_api_v1_missions__mission_id__transition_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionMissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transition Mission",
        "tags": [
          "missions"
        ]
      }
    },
...
  }
}